implemented QuizSQL, added TestCases
1. added quizzes table building code
2. implemented QuizSQL
3. added tests for QuizSQL (100 % coverage, except error(catch) cases)

+++ Quiz-Website/src/main/java/Objects/Quiz.java
@@ -17,7 +17,7 @@ public class Quiz {
 
 
 
-    public Quiz(int quizId, long creatorId, String quizName, String description, Date createTime, boolean isDraft, boolean isPractice) {
+    public Quiz(long quizId, long creatorId, String quizName, String description, Date createTime, boolean isDraft, boolean isPractice) {
         this.quizId = quizId;
         this.creatorId = creatorId;
         this.quizName = quizName;
@@ -30,26 +30,18 @@ public class Quiz {
     public long getQuizId(){
         return quizId;
     }
-
-    public long getCreatorId(){
-        return creatorId;
-    }
-
+    public long getCreatorId(){ return creatorId;}
     public String getQuizName(){
         return quizName;
     }
-
     public String getDescription(){
         return description;
     }
-
     public Date getCreateTime(){
         return createTime;
     }
-
     public boolean isDraft(){
         return isDraft;
     }
-
     public boolean isPractice() { return isPractice; }
 }
